QuizApp/quizapp/requirements.txt
appdirs==1.4.4
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
distlib==0.3.1
Django==3.0.2
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
mysqlclient==1.4.6
pipenv==2020.11.15
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
virtualenv==20.1.0
virtualenv-clone==0.5.3
